UI and GameManager
I updated the UI again. button clicking practice on the song list page will now put you on the instrument select page, and selecting an instrument will take you to the recording page. And the users choices are stored in the GameManager. Scaling is still an issue and there is no backing out functions as of now.