UI update
After meeting with Josh. We decided to have to user choose a player. On entering the game the user will either select an avatar or load previous player avatar. The user is then taken to the song selection page which currently has not changed. After selecting a song the user is taken to the record page. On the record page the players avatar will be displayed holding the selected instrument. The instrument will be selected with the buttons at the bottom. Pressing an instrument button  while the instrument is the SelectedInstrument will toggle the audio clip being played between teacher, student, and muted.