Unity Version Update
When installing the new version of unity you may need to update the Multiplayer HLAPI in the package manager. This can be done by selecting Window>Package Manager, then selecting the Multiplayer HLAPI on the left hand side, and finally clicking update in the bottom right corner of the window.